requirements.txt
speechrecognition>=3.10.0
pyttsx3>=2.90
requests>=2.31.0
numpy>=1.24.0
pyaudio>=0.2.11
comtypes>=1.1.14
pywin32>=306
urllib3>=2.0.0
typing-extensions>=4.7.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.0.0
